refactor: use synchronous file system APIs and other nits
* Move from async to sync file operations. This is done because
for a small set of files/operations, sync is actually faster than
async in Node. Also the `exists` API is deprecated, but the
`existsSync` API is not
* Move to single-quote imports where needed
