+++ examples/podinfo/examples/app-example.ts
@@ -1,7 +1,7 @@
-import { Construct } from "@aws-cdk/core";
-import { App, Chart } from "cdk8s";
-import { Deployment, Service } from "../lib";
-import { PodinfoContainer } from "./podinfo";
+import { Construct } from '@aws-cdk/core';
+import { App, Chart } from 'cdk8s';
+import { Deployment, Service } from '../lib';
+import { PodinfoContainer } from './podinfo';
 
 class MyChart extends Chart {
   constructor(scope: Construct, id: string) {
