+++ examples/web-service/web-service.ts
@@ -1,9 +1,9 @@
 import { Construct } from '@aws-cdk/core';
-import { Service, IntOrString, Deployment } from '.././imports/k8s';
+import { Service, IntOrString, Deployment } from '../imports/k8s';
 
 export interface WebServiceOptions {
   /** The Docker image to use for this service. */
-  readonly image: string; // docker image to use for this service
+  readonly image: string;
 
   /**
    * Number of replicas.
